# src/Venture.IdentityManagement.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Venture.IdentityManagement",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Role/{roleId}": {
      "get": {
        "tags": [
          "Role"
        ],
        "summary": "Get Role by id.",
        "description": "Get Role by id.",
        "operationId": "getRoleById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetRoleByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Role"
        ],
        "summary": "Delete Role.",
        "description": "Delete Role.",
        "operationId": "removeRole",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveRoleResponse"
            }
          }
        }
      }
    },
    "/api/Role": {
      "get": {
        "tags": [
          "Role"
        ],
        "summary": "Get Roles.",
        "description": "Get Roles.",
        "operationId": "getRoles",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetRolesResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Role"
        ],
        "summary": "Create Role.",
        "description": "Create Role.",
        "operationId": "createRole",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateRoleRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateRoleResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Role"
        ],
        "summary": "Update Role.",
        "description": "Update Role.",
        "operationId": "updateRole",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateRoleRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateRoleResponse"
            }
          }
        }
      }
    },
    "/api/Role/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Role"
        ],
        "summary": "Get Role Page.",
        "description": "Get Role Page.",
        "operationId": "getRolesPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetRolesPageResponse"
            }
          }
        }
      }
    },
    "/api/Tenant/{tenantId}": {
      "get": {
        "tags": [
          "Tenant"
        ],
        "summary": "Get Tenant by id.",
        "description": "Get Tenant by id.",
        "operationId": "getTenantById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetTenantByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tenant"
        ],
        "summary": "Delete Tenant.",
        "description": "Delete Tenant.",
        "operationId": "removeTenant",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "tenantId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveTenantResponse"
            }
          }
        }
      }
    },
    "/api/Tenant": {
      "get": {
        "tags": [
          "Tenant"
        ],
        "summary": "Get Tenants.",
        "description": "Get Tenants.",
        "operationId": "getTenants",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetTenantsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tenant"
        ],
        "summary": "Create Tenant.",
        "description": "Create Tenant.",
        "operationId": "createTenant",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateTenantRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateTenantResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tenant"
        ],
        "summary": "Update Tenant.",
        "description": "Update Tenant.",
        "operationId": "updateTenant",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateTenantRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateTenantResponse"
            }
          }
        }
      }
    },
    "/api/Tenant/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Tenant"
        ],
        "summary": "Get Tenant Page.",
        "description": "Get Tenant Page.",
        "operationId": "getTenantsPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetTenantsPageResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateRoleRequest": {
      "type": "object",
      "properties": {
        "role": {
          "$ref": "#/definitions/RoleDto"
        }
      }
    },
    "CreateRoleResponse": {
      "type": "object",
      "properties": {
        "role": {
          "$ref": "#/definitions/RoleDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "CreateTenantRequest": {
      "type": "object",
      "properties": {
        "tenant": {
          "$ref": "#/definitions/TenantDto"
        }
      }
    },
    "CreateTenantResponse": {
      "type": "object",
      "properties": {
        "tenant": {
          "$ref": "#/definitions/TenantDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "GetRoleByIdResponse": {
      "type": "object",
      "properties": {
        "role": {
          "$ref": "#/definitions/RoleDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetRolesPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RoleDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetRolesResponse": {
      "type": "object",
      "properties": {
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RoleDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetTenantByIdResponse": {
      "type": "object",
      "properties": {
        "tenant": {
          "$ref": "#/definitions/TenantDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetTenantsPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TenantDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetTenantsResponse": {
      "type": "object",
      "properties": {
        "tenants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TenantDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
    "RemoveRoleResponse": {
      "type": "object",
      "properties": {
        "role": {
          "$ref": "#/definitions/RoleDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "RemoveTenantResponse": {
      "type": "object",
      "properties": {
        "tenant": {
          "$ref": "#/definitions/TenantDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "RoleDto": {
      "type": "object",
      "properties": {
        "roleId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "TenantDto": {
      "type": "object",
      "properties": {
        "tenantId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "UpdateRoleRequest": {
      "type": "object",
      "properties": {
        "role": {
          "$ref": "#/definitions/RoleDto"
        }
      }
    },
    "UpdateRoleResponse": {
      "type": "object",
      "properties": {
        "role": {
          "$ref": "#/definitions/RoleDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "UpdateTenantRequest": {
      "type": "object",
      "properties": {
        "tenant": {
          "$ref": "#/definitions/TenantDto"
        }
      }
    },
    "UpdateTenantResponse": {
      "type": "object",
      "properties": {
        "tenant": {
          "$ref": "#/definitions/TenantDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}